Check if user exists

+++ backendstuff/usagescripts/destroyuser.sh
@@ -10,6 +10,12 @@ else
   exit 1
 fi
 
+# Check that user exists
+if [! id "$username" &>/dev/null]; then
+  echo "$date - Tried to destroy nonexistent user $username" >> log.txt
+  exit 1
+fi
+
 userdel -r $1
 
 # Log
